backend/requirements.txt
Flask
Flask-Cors
oracledb
python-dotenv
requests
bcrypt
PyJWT
Flask-APScheduler
Flask-SocketIO
eventlet
